+++ src/connectors/sql/schema/column/mod.rs
@@ -63,7 +63,7 @@ impl SQLColumn {
 impl ToSQLString for SQLColumn {
     fn to_string(&self, dialect: SQLDialect) -> String {
         let name = &self.name;
-        let t = self.r#type.to_string(dialect.clone());
+        let t = self.r#type.to_string(dialect);
         let not_null = if self.not_null { " NOT NULL" } else { " NULL" };
         let primary = if self.primary_key { " PRIMARY KEY" } else { "" };
         let default = if self.default.is_some() { " DEFAULT ".to_owned() + self.default.as_ref().unwrap().as_str() } else { "".to_owned() };